new floating ip scripts
